{"schema":"decodex/commit/1","summary":"Split search ranking payload helpers into a child module after strict validation.","authority":"manual"}
